acpi: rsdp: add SBBR compliance tests
Server Base Boot Requirements (SBBR) specification is
intended for SBSA- compliant 64-bit ARMv8 servers.
It defines the base firmware requirements for out-of-box support
of any ARM SBSA-compatible Operating System or hypervisor. The requirements
in this specification are expected to be minimal yet complete for
booting a multi-core ARMv8 server platform, while leaving plenty of
room for OEM or ODM innovations and design details.
For more information, download the SBBR specification here:
http://infocenter.arm.com/help/index.jsp?topic=/com.arm.doc.den0044b/index.html

This change introduces test cases as per SBBR specification to rsdp acpi.

Signed-off-by: Supreeth Venkatesh <supreeth.venkatesh@arm.com>
Signed-off-by: Sakar Arora <Sakar.Arora@arm.com>